feat(llvmgen): add type, global, and func buffers before writing to Writer
By writing global variables, type definitions, and function definitions to three
separate buffers, we can later write them to the Writer in the correct order all
at once. Previously it was difficult to emit type definitions before their use
when the code generator emitted code as it traversed the AST. Now, the type
definitions are guaranteed to be printed before they are referenced in function
bodies.